// ModemAgent/web-app/src/types.ts
export interface CarrierComponent {
  label: string // "PCC", "SCC0", "SCC1", etc.
  band: string // "B8", "n78"
  pci: number
  earfcn: number
  bandwidth: string // "10 MHz"
  freq?: number // MHz, calculated from EARFCN
  rsrp?: number
  rsrq?: number
  sinr?: number
  rssi?: number
  ul_configured?: boolean
  active?: boolean
}

export interface SignalInfo {
  type?: string
  carrier?: string
  signal_bars?: number
  cell_id?: string
  lte_carriers: CarrierComponent[]
  nr_carriers: CarrierComponent[]
  net_select?: string
  lte_band_lock?: number[]
  nr_band_lock?: number[]
  raw_lte_band_lock?: string
  raw_nr_band_lock?: string
  rsrp?: number
  band?: string
}
// ...
export interface ModemCapabilities {
  network_modes: { value: string; label: string }[]
  lte_bands: number[]
  nr_sa_bands: number[]
  nr_nsa_band_lock_supported: boolean
}